chore(hydro_lang): move backtrace test to separate module for stable line numbers
Will reduce diff noise whenever modifying `Stream`.
