+++ hydro_lang/src/live_collections/stream/mod.rs
@@ -2505,6 +2505,8 @@ mod tests {
     use crate::compile::builder::FlowBuilder;
     use crate::location::Location;
 
+    mod backtrace_chained_ops;
+
     struct P1 {}
     struct P2 {}
 
@@ -2576,42 +2578,5 @@ mod tests {
         deployment.start().await.unwrap();
 
         assert_eq!(external_out.next().await.unwrap(), 1);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn backtrace_chained_ops() {
-        use crate::compile::ir::HydroRoot;
-
-        let flow = FlowBuilder::new();
-        let node = flow.process::<()>();
-
-        node.source_iter(q!([123])).for_each(q!(|_| {}));
-
-        let finalized: crate::compile::built::BuiltFlow<'_> = flow.finalize();
-
-        let source_meta = if let HydroRoot::ForEach { input, .. } = &finalized.ir()[0] {
-            use crate::compile::ir::HydroNode;
-
-            if let HydroNode::Unpersist { inner, .. } = input.as_ref() {
-                if let HydroNode::Persist { inner, .. } = inner.as_ref() {
-                    if let HydroNode::Source { metadata, .. } = inner.as_ref() {
-                        &metadata.op
-                    } else {
-                        panic!()
-                    }
-                } else {
-                    panic!()
-                }
-            } else {
-                panic!()
-            }
-        } else {
-            panic!()
-        };
-        let for_each_meta = finalized.ir()[0].op_metadata();
-
-        hydro_build_utils::assert_debug_snapshot!(source_meta.backtrace.elements());
-        hydro_build_utils::assert_debug_snapshot!(for_each_meta.backtrace.elements());
     }
 }
